サービス切り出しでリファクタリング
