+++ rails/app/services/student/analytics/task_completion.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Student
+  module Analytics
+    class TaskCompletion
+      def self.call(user)
+        new(user).call
+      end
+
+      def initialize(user)
+        @user = user
+      end
+
+      def call
+        task_stats = TaskStats.call(@user)
+
+        completed_count = task_stats.completed_count
+        total_count = task_stats.total_count
+        {
+          completed_count: completed_count,
+          total_count: total_count,
+          completion_rate: Calculator.completion_rate(completed_count, total_count)
+        }
+      end
+    end
+  end
+end
